+++ functions.php
@@ -38,3 +38,5 @@ require get_template_directory() . '/inc/news.php';
 require get_template_directory() . '/inc/performance.php';
 
 require get_template_directory() . '/inc/security.php';
+
+require get_template_directory() . '/inc/schema.php';